jav-508 add metrics publisher and move metrics to metric package
